+++ data/cocoatable.js
@@ -298,13 +298,11 @@ try	{
 		console.log(str);
 	}
 
-	CocoaTable.prototype.setupEtcButtonHandlers = function () {
-		var restoreButton = document.getElementById('restore-button');
+	function setupRestoreButton(self) {
+ 		var restoreButton = document.getElementById('restore-button');
 
-		var self = this;
-	    
 		restoreButton.addEventListener( 'click', function (ev) {
- 			if (confirm('restore setting?')) {
+ 			if (confirm('Restore Setting?')) {
 				// 既存のテーブルは削除
 				var tbody = document.querySelector('.cocoatable tbody');
 				removeChildAll(tbody);
@@ -320,9 +318,57 @@ try	{
 
 				self.initalize(defaultLinkformData);
 				self.updated();
+
+				// 通知
+				self.pushButton("restore");
  			}
 		}, true);
 	}
+
+	function setupClearButton(self) {
+ 		var button = document.getElementById('clear-button');
+
+		button.addEventListener( 'click', function (ev) {
+ 			if (confirm('Clear Setting?')) {
+				// 既存のテーブルは削除
+				var tbody = document.querySelector('.cocoatable tbody');
+				removeChildAll(tbody);
+
+				// 初期化
+				self.initalize([]);
+				self.updated();
+
+				// 通知
+				self.pushButton("clear");
+ 			}
+		}, true);
+	}
+
+    function setupImportButton(self) {
+		var button = document.getElementById('import-button');
+
+		button.addEventListener( 'click', function (ev) {
+			self.updated();
+			self.pushButton("import");
+		}, true);
+	}
+
+    function setupExportButton(self) {
+		var button = document.getElementById('export-button');
+
+		button.addEventListener( 'click', function (ev) {
+			self.updated();
+			self.pushButton("export");
+		}, true);
+	}
+
+	CocoaTable.prototype.setupEtcButtonHandlers = function () {
+		setupRestoreButton(this);
+		setupImportButton(this);
+		setupExportButton(this);
+		setupClearButton(this);
+	}
+
 	CocoaTable.Cell.prototype.element = function () {
 		return this._e;
 	}
@@ -488,4 +534,13 @@ try	{
 		if ( this._listener.onUpdated )
 			this._listener.onUpdated();
 	}
+	CocoaTable.prototype.pushButton = function (msg) {
+		if ( this._listener.onPushButton )
+			this._listener.onPushButton(msg);
+	}
+	CocoaTable.prototype.importData = function (data) {
+		this.initalize(data);
+		this.updated();
+	}
+