+++ tests/test-firelink_lib.js
@@ -38,11 +38,18 @@ function test_isgd(test) {
 function test_copyLinkAndNotify(test) {
   let linkdata = {text: "YAAOO!! JIPAN", title: "YAAOO!! JIPAN", url: "http://www.example.jp"};
   
-  fl.copyLinkAndNotify("DUMMY", "%text%\n%url%", linkdata);
+  fl.copyLinkAndNotify("DUMMY", "%text%\\n%url%", linkdata);
   test.assertEqual("YAAOO!! JIPAN\nhttp://www.example.jp", clipboard.get());
 
   fl.copyLinkAndNotify("DUMMY2", "<a href=\"%url%\">%text%</a>", linkdata);
   test.assertEqual("<a href=\"http://www.example.jp\">YAAOO!! JIPAN</a>", clipboard.get());
+}
+
+function test_createText(test) {
+  let linkform = "%text%\\n%url%";
+  let linkdata = {text: "aaa\nbbb\nccc", title: "", url: "http://www.example.jp"};
+  
+  test.assertEqual("aaa bbb ccc\nhttp://www.example.jp", fl.createText(linkform, linkdata));
 }
 
 exports.test_test_run = function(test) {
@@ -61,5 +68,49 @@ exports.test_test_run = function(test) {
   fl.changeLinkform("[[%text%|%url%]]");
   test.assertEqual(3, fl.currentIndex());
 
+  // 
+  test_createText(test);
+
+  // json2Text
+  {
+    let json = [
+      {name:"PlainText",   format:"%text%\\n%url%"},
+      {name:"HTML",        format:"<a href=\"%url%\">%text%</a>"},
+      {name:"Twitter",     format:"%text% %isgd%"},
+      {name:"TiddlyWiki",  format:"[[%text%|%url%]]"}
+    ];
+
+    let text =
+      "PlainText\t%text%\\n%url%\n" +
+      "HTML\t<a href=\"%url%\">%text%</a>\n" +
+      "Twitter\t%text% %isgd%\n" +
+      "TiddlyWiki\t[[%text%|%url%]]\n";
+
+    test.assertEqual(fl.json2Text(json), text);
+    test.assertEqual(JSON.stringify(fl.text2JSON(text)), JSON.stringify(json));
+  }
+
+  // patten2
+  {
+    let json = [
+      {name:"HTML",        format:"<a href=\"%url%\">%text%</a>"},
+      {name:"PlainText",   format:"%text%\\n%url%"},
+    ];
+
+    let text =
+      "HTML\t<a href=\"%url%\">%text%</a>\n" + 
+      "PlainText\t%text%\\n%url%\n";
+
+    test.assertEqual(JSON.stringify(fl.text2JSON(text)), JSON.stringify(json));
+  }
+
+  // test
+  test.assertEqual(fl.addPrefix("hoge", 0), "(1) hoge");
+  test.assertEqual(fl.addPrefix("hoge", 1), "(2) hoge");
+  test.assertEqual(fl.addPrefix("hoge", 2), "(3) hoge");
+  test.assertEqual(fl.addPrefix("hoge", 8), "(9) hoge");
+  test.assertEqual(fl.addPrefix("hoge", 9), "(0) hoge");
+  test.assertEqual(fl.addPrefix("hoge", 15), "hoge");
+  
 };