+++ tests/test-menu_lib.js
@@ -31,7 +31,3 @@ exports.test_test_run = function(test) {
   MenuLib.copyTabs(action, linkdata);
 };
 
-
-
-
-