+++ HISTORY.md
@@ -1,3 +1,17 @@
+# 0.9 2011/05/07
+
+* Import & Export
+  * Linkform setting share & backup.
+  * Import from clipboard
+  * Export to clipboard
+* Configuration screen redesign.
+* Add "Recommended Linkform" in Setting panel.
+* Add "Set %text% from Clipboard (Ctrl+X)".
+* "Direct Select", "0" key add.
+* The selected text to all line breaks converted to spaces.
+* Add "Clear Setting" button.
+* Typo, "All Tabs Space" -> "All Tabs Separate".
+
 # 0.8 2011/04/30
 
 * "All Tabs" command can be chosen linkform.